setup, CI: fix macos build, cleanup

+++ setup.py
@@ -21,6 +21,7 @@ import os
 import os.path
 import platform
 import shutil
+import subprocess
 import sys
 import typing as t
 from pathlib import Path
@@ -32,6 +33,10 @@ from setuptools.command.build_ext import build_ext
 
 __all__ = ["setup"]
 
+
+is_macos = platform.system() == "Darwin"
+is_windows = platform.system() == "Windows"
+dll_ext = ".dll" if is_windows else ".dylib" if is_macos else ".so"
 
 debug_c_args = {
     "unix": ["-Og", "-g"],
@@ -48,8 +53,8 @@ release_c_args = {
 debug_l_args: t.Dict[str, t.List[str]] = {}
 
 release_l_args = {
-    "unix": ["-s", "-Wl,--gc-sections", "-Wl,-rpath,$ORIGIN", "-L."],
-    "gcc": ["-s", "-Wl,--gc-sections", "-Wl,-rpath,$ORIGIN", "-L."],
+    "unix": ["-s", "-Wl,--gc-sections", "-Wl,-rpath,$ORIGIN"],
+    "gcc": ["-s", "-Wl,--gc-sections", "-Wl,-rpath,$ORIGIN"],
 }
 
 release_libraries = {
@@ -61,17 +66,38 @@ release_libraries = {
 c_args = release_c_args
 l_args = release_l_args
 libraries = release_libraries
+library_dirs = ["."]
 
-if platform.system() == "Darwin":
-    for tool in l_args:  # gc-sections not supported by llvm; DLL ext is .dylib
-        if "-Wl,--gc-sections" in l_args[tool]:
-            l_args[tool].remove("-Wl,--gc-sections")
+if is_macos:
+    # macOS (clang) uses some different compiler switches
+    c_args_replacements: t.List[t.Tuple[str, t.Union[str, None]]] = [
+        ("-s", None),
+    ]
+    l_args_replacements: t.List[t.Tuple[str, t.Union[str, None]]] = [
+        ("-Wl,--gc-sections", None),
+        ("-s", None),
+        ("-Wl,-rpath,$ORIGIN", "-Wl,-rpath,@loader_path"),
+    ]
+    for tool in c_args:
+        for orig, replacement in c_args_replacements:
+            if orig in c_args[tool]:
+                c_args[tool].remove(orig)
+                if replacement:
+                    c_args[tool].append(replacement)
+    for tool in l_args:
+        for orig, replacement in l_args_replacements:
+            if orig in l_args[tool]:
+                l_args[tool].remove(orig)
+                if replacement:
+                    l_args[tool].append(replacement)
     for tool in libraries:
         for i, lib in enumerate(libraries[tool]):
-            if lib.startswith(":") and lib.endswith(".so"):
+            # For some reason using a direct path doesn't work during link-time on macOS.
+            # See explanation in CustomCommand.run.
+            if lib.startswith(":SoMRandomizer.api."):
+                libraries[tool][i] = "SoMRandomizer-api"
+            elif lib.endswith(".so"):
                 libraries[tool][i] = lib[:-3] + ".dylib"
-
-dll_ext = ".dll" if platform.system() == "Windows" else ".dylib" if platform.system() == "Darwin" else ".so"
 
 
 def get_dotnet_os():
@@ -126,27 +152,64 @@ class CustomCommand(Command):
         if bdist.bdist_dir is not None:
             self.bdist_dir = Path(bdist.bdist_dir)
 
-    def run(self) -> None:
-        # NOTE: we need the DLL here so we can properly link it
-        import subprocess
+    @staticmethod
+    def dotnet_publish(rid: str) -> Path:
+        if not rid.replace("-", "").isalnum():
+            raise ValueError("Invalid rid")
 
-        # contrary to documentation, it appears that we have to explicitly set the RID
-        rid = "-".join((get_dotnet_os(), get_dotnet_arch()))
+        # contrary to documentation, it appears that we have to explicitly set the RID even for "native"
+        # also supplying -r <rid> fails for whatever reason
         with open("SecretOfManaRandomizer/SoMRandomizer.api/SoMRandomizer.api.csproj.user", "w") as f:
             f.write(f"<Project><PropertyGroup><RuntimeIdentifier>{rid}</RuntimeIdentifier></PropertyGroup></Project>")
-        # TODO: on macos, we want to build both x86_64 and arm64 and use lipo to get a universal2 dylib
+        dotnet = shutil.which("dotnet")
+        assert dotnet, "dotnet not found"
         subprocess.run(
-            "dotnet publish SecretOfManaRandomizer/SoMRandomizer.api/",
-            shell=True,
+            [dotnet, "publish", "SecretOfManaRandomizer/SoMRandomizer.api/"],
             check=True,
         )
-        native_lib = f"SoMRandomizer.api{dll_ext}"
-        native_lib_src = f"SecretOfManaRandomizer/SoMRandomizer.api/bin/Release/net10.0/native/{native_lib}"
-        print(f"copying {native_lib_src} -> {os.path.curdir} ({os.path.getsize(native_lib_src)})")
-        shutil.copy(native_lib_src, ".")
+        native_lib_src_name = f"SoMRandomizer.api{dll_ext}"
+        return Path(f"SecretOfManaRandomizer/SoMRandomizer.api/bin/Release/net10.0/native/{native_lib_src_name}")
+
+    def dotnet_publish_universal2(self) -> Path:
+        macos_build_dir = Path("build") / "macos"
+        macos_build_dir.mkdir(parents=True, exist_ok=True)
+        dotnet_os = get_dotnet_os()
+        arm64_src = self.dotnet_publish(f"{dotnet_os}-arm64")
+        arm64_lib = macos_build_dir / Path(arm64_src).name.replace(".dylib", ".arm64.dylib")
+        shutil.move(arm64_src, arm64_lib)
+        x64_src = self.dotnet_publish(f"{dotnet_os}-x64")
+        x64_lib = macos_build_dir / Path(arm64_src).name.replace(".dylib", ".x64.dylib")
+        shutil.move(x64_src, x64_lib)
+        universal_lib = macos_build_dir / Path(x64_src).name
+        lipo = shutil.which("lipo")
+        assert lipo, "lipo not found"
+        subprocess.run(
+            [lipo, "-output", str(universal_lib), "-create", str(arm64_lib), str(x64_lib)],
+            check=True,
+        )
+        return universal_lib
+
+    def run(self) -> None:
+        # NOTE: we need the DLL here so we can properly link it, also we want to package it further down
+        if is_macos:
+            native_lib_src = self.dotnet_publish_universal2()
+        else:
+            rid = "-".join((get_dotnet_os(), get_dotnet_arch()))
+            native_lib_src = self.dotnet_publish(rid)
+        native_lib = native_lib_src.name
+        print(f"copying {native_lib_src} -> {native_lib} ({os.path.getsize(native_lib_src)})")
+        shutil.copy(native_lib_src, native_lib)
+
+        if is_macos:
+            # for whatever reason it looks up the correct name during runtime, but can't link it during link time,
+            # so we copy the badly-named dll to the name specified above and then ship the badly-named one anyway.
+            temp_lib = f"lib{native_lib.replace('.api', '-api')}"
+            print(f"copying {native_lib} -> {temp_lib}")
+            shutil.copy(native_lib, temp_lib)
+
         if dll_ext == ".dll":
-            # also copy importlib. confusingly this is not called .dll.lib, but just .lib
-            import_lib_src = native_lib_src[:-4] + ".lib"
+            # also copy importlib on windows
+            import_lib_src = native_lib_src.with_suffix(".lib")
             print(f"copying {import_lib_src} -> {os.path.curdir} ({os.path.getsize(import_lib_src)})")
             shutil.copy(import_lib_src, ".")
 
@@ -181,6 +244,7 @@ class CustomBuildExt(build_ext):
         if c in libraries:
             for e in self.extensions:
                 e.libraries = libraries[c]
+                e.library_dirs = library_dirs
         return build_ext.build_extensions(self)
